refactor(search applications): rename engine folder
