+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_application/add_indices_logic.test.ts
