+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_application/field_icon.tsx
