+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_application/header_docs_action.tsx
