+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_application/search_application_content.tsx
