+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_application/search_preview/document_flyout.tsx
