+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_application/search_preview/field_value_cell.tsx
