+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_application/search_preview/search_preview_logic.ts
